# backend/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "backend, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Candidate/Create": {
      "post": {
        "tags": [
          "Candidate"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "FirstName",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "LastName",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Email",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Phone",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "CoverLetter",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "JobId",
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "formData",
            "name": "pdfFile",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Candidate/Get": {
      "get": {
        "tags": [
          "Candidate"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CandidateGetDto"
              }
            }
          }
        }
      }
    },
    "/api/Candidate/Get{id}": {
      "get": {
        "tags": [
          "Candidate"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CandidateGetDto"
            }
          }
        }
      }
    },
    "/api/Candidate/download/{url}": {
      "get": {
        "tags": [
          "Candidate"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "url",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Company/Create": {
      "post": {
        "tags": [
          "Company"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CompanyCreateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Company/Get": {
      "get": {
        "tags": [
          "Company"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CompanyGetDto"
              }
            }
          }
        }
      }
    },
    "/api/Company/Get/{id}": {
      "get": {
        "tags": [
          "Company"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CompanyGetDto"
            }
          }
        }
      }
    },
    "/api/Job/Create": {
      "post": {
        "tags": [
          "Job"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/JobCreateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Job/Get": {
      "get": {
        "tags": [
          "Job"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/JobGetDto"
              }
            }
          }
        }
      }
    },
    "/api/Job/Get{id}": {
      "get": {
        "tags": [
          "Job"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/JobGetDto"
            }
          }
        }
      }
    },
    "/WeatherForecast": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "operationId": "GetWeatherForecast",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/WeatherForecast"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "CandidateGetDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "coverLetter": {
          "type": "string"
        },
        "resumeUrl": {
          "type": "string"
        },
        "jobId": {
          "format": "int64",
          "type": "integer"
        },
        "jobTitle": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "CompanyCreateDto": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "size": {
          "$ref": "#/definitions/CompanySize"
        }
      },
      "additionalProperties": false
    },
    "CompanyGetDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "size": {
          "$ref": "#/definitions/CompanySize"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "CompanySize": {
      "enum": [
        "Small",
        "Medium",
        "Large"
      ],
      "type": "string"
    },
    "JobCreateDto": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "level": {
          "$ref": "#/definitions/JobLevel"
        },
        "companyId": {
          "format": "int64",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "JobGetDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "level": {
          "$ref": "#/definitions/JobLevel"
        },
        "companyId": {
          "format": "int64",
          "type": "integer"
        },
        "companyName": {
          "type": "string"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "JobLevel": {
      "enum": [
        "Intern",
        "Junior",
        "MidLevel",
        "Senior",
        "TeamLead",
        "Cto",
        "Architect"
      ],
      "type": "string"
    },
    "WeatherForecast": {
      "type": "object",
      "properties": {
        "date": {
          "format": "date",
          "type": "string"
        },
        "temperatureC": {
          "format": "int32",
          "type": "integer"
        },
        "temperatureF": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "summary": {
          "type": "string"
        }
      },
      "additionalProperties": false
    }
  }
}